moved tail types to cli/tail.go so definitions are localized to what they should be used for.
